+++ src/main/java/com/spring/koction/controller/item/ItemController.java
@@ -1,12 +1,12 @@
 package com.spring.koction.controller.item;
 
-import com.spring.koction.commons.FileUtils;
-import com.spring.koction.entity.Item;
-import com.spring.koction.entity.ItemFile;
-import com.spring.koction.service.item.ItemService;
+import java.io.IOException;
+import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.parameters.P;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.List;
+import com.spring.koction.commons.FileUtils;
+import com.spring.koction.entity.Item;
+import com.spring.koction.entity.ItemFile;
+import com.spring.koction.service.item.ItemService;
 
 @RestController
 @RequestMapping("/item")
@@ -37,18 +37,28 @@ public class ItemController {
 
 
 	@GetMapping("/registerItem")
-	public ModelAndView bidView() {
+	public ModelAndView registerItemView() {
 		ModelAndView mv = new ModelAndView();
+		mv.setViewName("/item/itemRegister.html");
 		return mv;
 	}
-
+	
 	@PostMapping("/registerItem")
-	public void insertBoard(HttpServletResponse httpServletResponse, Item item, HttpServletRequest httpServletRequest, MultipartHttpServletRequest multipartHttpServletRequest) throws IOException {
-		int itemNo = itemService.registerItem(item);
+	public ModelAndView registerItem(Item item, int term, HttpServletRequest request, MultipartHttpServletRequest multipartServletRequest) throws IOException {
+		item.setItemEnddate(item.getItemRegdate().plusDays(term));
+		int itemNo = itemService.regitserItem(item);//글등록 및 글 번호 반환
+		
 		FileUtils fileUtils = new FileUtils();
-		List<ItemFile> fileList = fileUtils.parseFileInfo(itemNo, httpServletRequest, multipartHttpServletRequest);
-		itemService.registerItemFileList(fileList);
-		httpServletResponse.sendRedirect("/item/myItem");
+		List<ItemFile> fileList = fileUtils.parseFileInfo(itemNo, request, multipartServletRequest);
+		
+		for(ItemFile itemFile : fileList) {
+			System.out.println(itemFile.getItem());
+		}
+		itemService.registerItemFile(fileList);
+		
+		ModelAndView mv = new ModelAndView();
+		mv.setViewName("/item/itemRegister.html");
+		return mv;
 	}
 	
 	@PostMapping("/updateItem")
@@ -56,7 +66,7 @@ public class ItemController {
 		itemService.updateItem(item);
 		FileUtils fileUtils = new FileUtils();
 		List<ItemFile> fileList = fileUtils.parseFileInfo(item.getItemNo(), httpServletRequest, multipartHttpServletRequest);
-		itemService.registerItemFileList(fileList);
+		itemService.registerItemFile(fileList);
 		httpServletResponse.sendRedirect("/item/myItem");
 	}
